+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file[^1].
 ### Added
 - item style period, current location
 
+### Fixed
+- newsletter sign-up to include marketing permissions
+
 ## [2.49.0] - 2022-03-25
 ### Changed
 - galleries list to come from a config file